Create EPI
Normally in denoising applications to find the quality of images we use quantitative metrics like PSNR,IEF,SSIM. When an impulse noise is available in a given image, during denoising the edges will be removed to a certain extent.To ensure that the resultant image after denoising preserves edges there is an index called Edge preservation Index(EPI) given by Farook Sattar et al in the paper "Image Enhancement based on a Nonlinear multi scale method", IEEE Trans on Image Processing,Vol.6, No.6, June 1997. The above coed was written using the equation available in the paper. It is more or less like correlation index ranging between 0 to 1